perf(ci): balance release E2E and reuse npm downloads
