+++ .github/workflows/openclaw-cross-os-release-checks-reusable.yml
@@ -921,6 +921,7 @@ jobs:
       # Windows Server 2025 with Node 24.19 can crash in libuv fs-event on RUNNER~1.
       # Keep only packaged-fresh on the last validated Node until an upstream fix ships.
       - name: Setup Node.js
+        id: node
         uses: actions/setup-node@820762786026740c76f36085b0efc47a31fe5020 # v7.0.0
         with:
           node-version: ${{ matrix.os_id == 'windows' && matrix.suite == 'packaged-fresh' && '24.15.0' || env.NODE_VERSION }}
@@ -932,6 +933,19 @@ jobs:
           package-manager-file: workflow/package.json
           lockfile-path: workflow/pnpm-lock.yaml
           cache-mode: off
+
+      # Lane homes and installed trees stay fresh. Only npm's integrity-checked
+      # download store survives; logs and npx installations never enter this cache.
+      - name: Restore npm downloads
+        id: npm_downloads
+        uses: actions/cache/restore@55cc8345863c7cc4c66a329aec7e433d2d1c52a9 # v6.1.0
+        with:
+          path: .cache/openclaw-cross-os-npm-cache/_cacache
+          enableCrossOsArchive: true
+          key: openclaw-cross-os-npm-v1-${{ runner.os }}-${{ runner.arch }}-${{ steps.node.outputs.node-version }}-${{ matrix.suite }}-${{ needs.prepare.outputs.candidate_sha256 }}-${{ needs.prepare.outputs.baseline_sha256 }}
+          restore-keys: |
+            openclaw-cross-os-npm-v1-${{ runner.os }}-${{ runner.arch }}-${{ steps.node.outputs.node-version }}-${{ matrix.suite }}-
+            openclaw-cross-os-npm-v1-seed-
 
       - name: Download candidate artifact
         id: download_candidate
@@ -1044,6 +1058,7 @@ jobs:
       - name: Run cross-OS release checks
         shell: bash
         env:
+          NPM_CONFIG_CACHE: ${{ github.workspace }}/.cache/openclaw-cross-os-npm-cache
           OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
           ANTHROPIC_API_KEY: ${{ secrets.ANTHROPIC_API_KEY }}
           MINIMAX_API_KEY: ${{ secrets.MINIMAX_API_KEY }}
@@ -1113,6 +1128,15 @@ jobs:
 
           echo "::warning::retrying Windows release checks after the outer process exited 127 after the dashboard probe and before summary generation"
           run_cross_os_release_checks
+
+      - name: Save npm downloads
+        if: github.repository == 'openclaw/openclaw' && github.event_name == 'workflow_dispatch' && steps.npm_downloads.outputs.cache-hit != 'true'
+        continue-on-error: true
+        uses: actions/cache/save@55cc8345863c7cc4c66a329aec7e433d2d1c52a9 # v6.1.0
+        with:
+          path: .cache/openclaw-cross-os-npm-cache/_cacache
+          enableCrossOsArchive: true
+          key: ${{ steps.npm_downloads.outputs.cache-primary-key }}
 
       - name: Summarize release checks
         if: always()
