+++ test/scripts/openclaw-cross-os-release-workflow.test.ts
@@ -80,6 +80,32 @@ describe("cross-OS release checks workflow", () => {
     );
   });
 
+  it("reuses npm downloads across isolated lane homes without caching installed state", () => {
+    const consumer = job(readWorkflow(WORKFLOW_PATH), "cross_os_release_checks");
+    const run = step(consumer, "Run cross-OS release checks");
+    const restore = step(consumer, "Restore npm downloads");
+    const save = step(consumer, "Save npm downloads");
+    const cacheRoot = run.env?.NPM_CONFIG_CACHE;
+
+    expect(cacheRoot).toBe("${{ github.workspace }}/.cache/openclaw-cross-os-npm-cache");
+    expect(restore.with?.path).toBe(".cache/openclaw-cross-os-npm-cache/_cacache");
+    expect(save.with?.path).toBe(restore.with?.path);
+    expect(restore.with?.enableCrossOsArchive).toBe(true);
+    expect(save.with?.enableCrossOsArchive).toBe(true);
+    expect(restore.with?.["restore-keys"]).toContain("openclaw-cross-os-npm-v1-seed-\n");
+    expect(save.with?.key).toBe("${{ steps.npm_downloads.outputs.cache-primary-key }}");
+    expect(save.if).toBe(
+      "github.repository == 'openclaw/openclaw' && github.event_name == 'workflow_dispatch' && steps.npm_downloads.outputs.cache-hit != 'true'",
+    );
+    expect(step(consumer, "Setup Node.js").id).toBe("node");
+    expect(restore.with?.key).toBe(
+      "openclaw-cross-os-npm-v1-${{ runner.os }}-${{ runner.arch }}-${{ steps.node.outputs.node-version }}-${{ matrix.suite }}-${{ needs.prepare.outputs.candidate_sha256 }}-${{ needs.prepare.outputs.baseline_sha256 }}",
+    );
+    const steps = consumer.steps!;
+    expect(steps.indexOf(restore)).toBeLessThan(steps.indexOf(run));
+    expect(steps.indexOf(save)).toBeGreaterThan(steps.indexOf(run));
+  });
+
   it("retries only an interrupted Windows dashboard probe", () => {
     const workflow = readWorkflow(WORKFLOW_PATH);
     const consumer = job(workflow, "cross_os_release_checks");
