+++ test/vitest-ui-e2e-config.test.ts
@@ -47,6 +47,7 @@ function timingFile(fileSeconds: Record<string, number>, perFileOverheadSeconds 
     source: "fixture measurements",
     uiE2e: { fileSeconds, perFileOverheadSeconds },
     compactGroupSeconds: { blacksmith: {}, github: {} },
+    repoE2eFileSeconds: {},
   });
 }
